# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for watching crypto prices",
        "title": "Crypto Price API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/currency/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coins"
                ],
                "summary": "Add coin to watchlist",
                "parameters": [
                    {
                        "description": "Add coin request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_http_handlers_coin_addCoin.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "No content (only status)"
                    },
                    "400": {
                        "description": "Input error",
                        "schema": {
                            "$ref": "#/definitions/crypto-price-service_internal_delivery_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/crypto-price-service_internal_delivery_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/currency/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coins"
                ],
                "summary": "get coins watchlist",
                "responses": {
                    "200": {
                        "description": "Watchable coins",
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_http_handlers_coin_watchlist.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/crypto-price-service_internal_delivery_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/currency/price": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prices"
                ],
                "summary": "get closest coin price",
                "parameters": [
                    {
                        "description": "get closest price for coin",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_http_handlers_prices_closestToTimestamp.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ближайшая цена монеты",
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_http_handlers_prices_closestToTimestamp.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Input error",
                        "schema": {
                            "$ref": "#/definitions/crypto-price-service_internal_delivery_http_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Price does not exists yet",
                        "schema": {
                            "$ref": "#/definitions/crypto-price-service_internal_delivery_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/crypto-price-service_internal_delivery_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/currency/prices": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prices"
                ],
                "summary": "get all prices for coin",
                "parameters": [
                    {
                        "description": "get all prices for coin",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_http_handlers_prices_AllForCoin.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_http_handlers_prices_AllForCoin.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Coin doew not exists",
                        "schema": {
                            "$ref": "#/definitions/crypto-price-service_internal_delivery_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/crypto-price-service_internal_delivery_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/currency/remove": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coins"
                ],
                "summary": "remove coin from watchlist",
                "parameters": [
                    {
                        "description": "remove coin request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_http_handlers_coin_removeCoin.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No content (only status)"
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/crypto-price-service_internal_delivery_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "crypto-price-service_internal_delivery_http_dto.ErrorBody": {
            "type": "object",
            "properties": {
                "cause": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "detail": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "crypto-price-service_internal_delivery_http_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/crypto-price-service_internal_delivery_http_dto.ErrorBody"
                }
            }
        },
        "internal_delivery_http_handlers_coin_addCoin.Request": {
            "type": "object",
            "required": [
                "name",
                "symbol"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "internal_delivery_http_handlers_coin_removeCoin.Request": {
            "type": "object",
            "properties": {
                "Symbol": {
                    "type": "string"
                }
            }
        },
        "internal_delivery_http_handlers_coin_watchlist.Coin": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "internal_delivery_http_handlers_coin_watchlist.SuccessResponse": {
            "type": "object",
            "properties": {
                "coins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_delivery_http_handlers_coin_watchlist.Coin"
                    }
                }
            }
        },
        "internal_delivery_http_handlers_prices_AllForCoin.CoinPrice": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "internal_delivery_http_handlers_prices_AllForCoin.Request": {
            "type": "object",
            "properties": {
                "symbol": {
                    "type": "string"
                }
            }
        },
        "internal_delivery_http_handlers_prices_AllForCoin.SuccessResponse": {
            "type": "object",
            "properties": {
                "prices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_delivery_http_handlers_prices_AllForCoin.CoinPrice"
                    }
                }
            }
        },
        "internal_delivery_http_handlers_prices_closestToTimestamp.Request": {
            "type": "object",
            "properties": {
                "symbol": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "internal_delivery_http_handlers_prices_closestToTimestamp.SuccessResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                }
            }
        }
    }
}